flask analyzer api 설정. 에러 코드 추가
