+++ pages/api/get/comment.ts
@@ -8,6 +8,8 @@ const apiKey = process.env.YOUTUBE_API_KEY;
 export default async function handler(
     req: NextApiRequest, res: NextApiResponse
 ){
+    if(req.method !== "GET") return res.status(405).json({message : "Not Allowed Method"})
+
     const { videoId } = req.query;
 
     // videoId 유효성 검사
